% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{afinn}
\alias{afinn}
\title{afinn sentiments}
\usage{
data(afinn)
}
\value{
A tibble
}
\description{
Sentiments for afinn data. We've downloaded these, because
the get_sentiments function is interactive for first
time users. This makes it fail the CI
}
\examples{
data(afinn)
}
\keyword{datasets}
