% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{beer_tidy}
\alias{beer_tidy}
\title{Tidied Beer information}
\usage{
data(beer_tidy)
}
\value{
A data frame
}
\description{
Alcohol % and color of different beers.
Color is ranked from 1-50, 50 being black.
}
\examples{
data(beer_tidy)
}
\keyword{datasets}
