% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{bing}
\alias{bing}
\title{bing sentiments}
\usage{
data(bing)
}
\value{
A tibble
}
\description{
Sentiments for bing data. We've downloaded these, because
the get_sentiments function is interactive for first
time users. This makes it fail the CI
}
\examples{
data(bing)
}
\keyword{datasets}
