% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{cocktails}
\alias{cocktails}
\title{Ingredients of classic cocktails}
\usage{
data(cocktails)
}
\value{
A tibble
}
\description{
Ingredients of classic cocktails as an R list
}
\examples{
data(cocktails)
}
\keyword{datasets}
