% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{dickens}
\alias{dickens}
\title{Charles Dickens Books}
\usage{
data(dickens)
}
\value{
A data frame
}
\description{
Text from the Charles Dickens classics, Oliver Twist and David Copperfield.
}
\examples{
data(dickens)
}
\keyword{datasets}
