% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{files}
\alias{files}
\title{File names for Stringr practice}
\usage{
data(files)
}
\value{
A vector
}
\description{
Taken from repurrrsive package, but cleaned up.
}
\examples{
data(files)
}
\keyword{Vectors}
