% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{gapl}
\alias{gapl}
\title{Information of 142 countries}
\usage{
data(gapl)
}
\value{
A list
}
\description{
A recursive list containing info on 142 countries.
Stats taken from the gapminder data in the repurrrsive package
}
\examples{
data(gapl)
}
\keyword{lists}
