% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_happiness.R
\name{get_happiness}
\alias{get_happiness}
\title{Create a local copy of the happiness data}
\usage{
get_happiness()
}
\value{
logical TRUE if the operation was successful
}
\description{
This function will attempt to create a local copy of the happiness data sets
for practical 2.
}
