% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{happiness}
\alias{happiness}
\title{World Happiness}
\usage{
data(happinesss)
}
\value{
A list
}
\description{
World happiness information for 146 countries from around the world
}
\examples{
data(happiness)
}
\keyword{lists}
