% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{ingredients}
\alias{ingredients}
\title{Ingredients on 6 classic sandwiches}
\usage{
data(ingredients)
}
\value{
A list
}
\description{
Ingredients of 6 clasic sandwiches, as a lists
}
\examples{
data(ingredients)
}
\keyword{lists}
