% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{names}
\alias{names}
\title{Messed up names}
\usage{
data(dickens)
}
\value{
A data frame
}
\description{
We asked children with 20 different names to type their name.
}
\examples{
data(names)
}
\keyword{datasets}
