% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{okcupid}
\alias{okcupid}
\title{OK Cupid data set}
\usage{
data(okcupid)
}
\value{
A tibble
}
\description{
Data from ok cupid see https://github.com/rudeboybert/JSE_OkCupid
}
\examples{
data(okcupid)
}
\keyword{datasets}
