% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{planet}
\alias{planet}
\title{Star Wars Planets and Their Information}
\usage{
data(planet)
}
\value{
A tibble or data frame
}
\description{
Taken from repurrrsive package, but cleaned up.
}
\examples{
data(planet)
}
\keyword{datasets}
